Fixed missing testing infrastructure, seperated out custom hook
