+++ explorer/setupJest.js
@@ -1,0 +1,2 @@
+//setupJest.js or similar file
+require('jest-fetch-mock').enableMocks()
